Major restructuring of folders. Major rebuild of win checker.
Apparently Java doesn't like the folder structure I had. This one runs
just fine, though. As for the win checker, I'm breaking it up because it
was too long. I'm also using outerLoop and innerLoop variables instead
of columns and rows because Horizontal checking wasn't feasible without
a separate loop and board rotation turned out to be too much work for
the benefit it conferred.