Small changes to win checker.
I'm thinking of rebuilding it to be a lot simpler. Instead of counting
up and checking whether a win is possible in only one direction at a
time, I could just iterate through rows and columns, top-left to
bottom-right (or, 3,3 from the bottom-right), checking right, down,
descending, and ascending run possibilities as I go. It'd be one table
read and roughly as many square reads.