First fully functional check-in.
Some refactoring is *very* possible, but overall I'm pretty happy with
it. The tie detection isn't what I want, because I'd like to be able to
iterate through the board and see if there are any winning length lines
possible for any of the players. I could do it, but it would take longer
than I want to spend on this project.

I consider Level 1 complete.