Moderate updates (plus one big one)
Added height and width as methods on Board

Extended Utilities class

Separated Game model from Game controller

Updated BoardController not to store board dimensions (because they're
properties of the board, not the controller)

Added much lighter win checker. Now we check a single move to see if
it's a winner.

Renamed BoardIterator to BoardChecker and rewrote it to use a List and a
loop to check in each direction. If it were possible to have more
directions, we could now check them.

Note: It would be better to start at the bottom than the top, since the top rows are rarely filled and never in the early game.
Note: On a very large board, we could ignore a lot of the bottom rows. For example, if the lowest open row in any column were 10, we could safely ignore everything below 6, since the lowest rows have been checked and found not to include winning lines. This principle also applies to columns.