Major refactor + AI V1
I did a major refactor and killed a bunch of functions. I decided to
focus on the AI component and fix things as breaks.

We now have an AI player that will win or block when possible and play
random columns otherwise. The threat list I developed can easily be
extended to a moderate AI that will play to extend or block
lower-severity threats. An advanced AI is too much for me to integrate
into this project, given my personal time and energy constraints, but I
wrote the objects and large functions in such a way that it can be
accommodated without much structural work.

I'll still be checking in some refactors and formatting changes, and I
still need to build out the game setup interface for multiple AI players
on custom boards and win conditions, but the structure is all there.

I consider Level 2 complete with this checkin.